Diplom/Diplom_2/requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.3.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1

requests~=2.32.4
faker~=37.5.3